django-verdant/requirements.txt
Django==1.8.5
wagtail==1.2rc1
django-auth-ldap==1.2.6
Embedly==0.5.0
elasticsearch==1.4.0
psycopg2==2.6.1
stripe==1.9.8
tweepy==2.1
django-redis==3.8.2
django-widget-tweaks==1.3
humanize==0.5
python-dateutil==1.5
celery==3.1.17
django-celery==3.1.16
simplejson==3.6.5
django-modelcluster==1.0
django-compressor==1.5

# Developer requirements
django-debug-toolbar==0.11.0
django-debug-toolbar-template-timings==0.5.3
Fabric==1.6.1
Sphinx==1.2
sphinx-rtd-theme==0.1.5

# Importer requirements
html2text==3.200.3
Markdown==2.3.1
